Refactor status bar managing into a separate class. Add project title to status bar menu
